requirements.txt
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
Django==4.2.8
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-jsonapi==7.0.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
inflection==0.5.1
python-dotenv==1.0.1
sniffio==1.3.1
sqlparse==0.5.1